Updated Makefile and cachewrapper
Renamed cachewrapper.c and changed the macros to work with new
directory names. Changed the makefile to compile with cachewrapper.c
linked as well.
